Collection and publication of some metrics inside and outside the application

+++ src/Tolerance/Bridge/Symfony/MessageProfile/StoresRequestProfileWhenKernelTerminates.php
@@ -65,7 +65,7 @@ final class StoresRequestProfileWhenKernelTerminates
     private function generateTiming()
     {
         $start = array_key_exists('REQUEST_TIME_FLOAT', $_SERVER) ?
-            \DateTime::createFromFormat('U.u', (double) $_SERVER['REQUEST_TIME_FLOAT']) :
+            \DateTime::createFromFormat('U.u', (float) $_SERVER['REQUEST_TIME_FLOAT']) :
             (array_key_exists('REQUEST_TIME', $_SERVER) ?
                 \DateTime::createFromFormat('U', (int) $_SERVER['REQUEST_TIME']) :
                 new \DateTime()
